Updated in-depth tutorial to zend-mvc v3
Completely reviewed, and, in many cases, rewrote the in-depth tutorial.

In particular, this moves from the mapper + DB abstraction approach to a
CQRS-inspired approach with a single level of interfaces. This new approach
simplifies much of the code and explanations, allowing the ability to focus on
usage of the various components.

All code samples were pulled directly from a working application developed in
tandem with the tutorial.