Added hints for running PHPUnit on Windows.
If trying to run PHPUnit on Windows as mentioned in the tutorial it doesn't work, you also can't run PHPUnit globally without further setup. However wrapping the command in double quotes fixes this and runs PHPUnit from the composer bin directory.

This may be obvious to some who are used to the command line however as this is part of the Getting Started tutorial I think it is worth mentioning.