use AdapterInterface instead of Adapter
I was working through the tutorial (as a ZF newbie) and found I had to replace Adapter with AdapterInterface; otherwise a fatal error was thrown previewing the `/album/` page in the browser.
Still, Eclipse tells me: `AlbumTableGateway cannot be resolved to a type` but PHP does not report any issue in this regard. Well, there is no class `AlbumTableGateway` but obviously `::class` is allowed to be used on non-existing classes so Eclipse is wrong here displaying an error instead of a warning.